third_party/WebKit/LayoutTests/editing/execCommand/indent-paragraphs.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
      '^<div>',
        '<div>',
          '<span style="background-color: red;">',
            'Hello <img src="http://"> world',
          '</span>',
        '</div>',
        '<div>Hello <img src="http://"> world</span></div>',
        'Hello <img src="http://"> world',
      '</div>|',
    '</div>',
  ].join(''),
  'indent',
  [
    '<div contenteditable>',
      '<blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">',
        '<div>',
          '<div>',
            '<span style="background-color: red;">',
              '^Hello <img src="http://"> world',
            '</span>',
          '</div>',
        '</div>',
        '<div>',
          '<div>Hello <img src="http://"> world</div>',
        '</div>',
        '<div>Hello <img src="http://"> world|</div>',
      '</blockquote>',
    '</div>',
  ].join('')),
  'indenting three paragraphs with different hierarchies');
</script>
